All out of creative summaries. YES I ONLY HAD TWO
I don't want to keep putting "updated" though. Like, that's pretty
obvious BECAUSE I'M COMMITTING CHANGES.

And I've clearly been working on this too long today.

Scenes 1 and 2 are functional with player choice actually making a
difference; askQuestion has been implemented for all four possible
number of replies; radio buttons are hidden during the boy's dialogue so
that it's more obvious that the player needs to reply; player's response
is printed after a dialogue option is chosen; "next" and "reply" buttons
are disabled when not needed, because we can't trust users; added
boolean isQuestion to allow for user input; fixed typos; reworded some
sentences; the dialogue speed is not permanent (it seems too fast), I
just needed it to move faster while I was testing.

I'm so pleased with the progress we made today.