The game has graphics!
DONE
Images now appear when the image name and extension (example.png) is
passed to changePicture(String); added line wraps on the word so that
all dialogue is on screen; added a space to the start of each sentence
so that it doesn't look so weird; set frame size to account for large
image icon; added currently finished scene images to gameImage folder.

TO DO
Images need to be added in the right places in the right scenes; scene 4
needs to be added; radio buttons need to be deselected between choices;
reply button to be disabled if none of the choices are selected(?); set
pauses before and after, sometimes during scenes; see if some kind of
transition is possible.